+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Expression calculator changelog
 
 ## [Unreleased]
 #### Added
-
+- Frontend template app based on React, Vite, TypeScript (#40)
 
 #### Changed
 - Repository structure updated to include frontend code (#39)